Finally fix issue with MacOS
scikit-build-core was calling:

cmake --install [build_dir] --prefix [install_prefix] --strip

This last `--strip` was the culprit, since we were not using this
ANYWHERE else. The problem is that nrnivmodl is loaded dynamically, and
CMake is not aware of this, so on MacOS it seems to have stripped the
libraries of the `nmodl_reg` symbol.