Experiment with replacing `nrnivmodl`
The `nrnivmodl` script is...complicated. With the move to building mod
files using pure CMake, it makes sense to simplify it to the maximum
extent possible. This commit makes the following changes to how
`nrnivmodl` works now:

- there is a cmake file, `nrnivmodl.cmake`, which builds all of the mod
  files. The list of mod files is passed from the bash script as a cmake
definition, so that everything is delegated to cmake as soon as possible
- the above file is currently `configure_file`-able; maybe an API
  version would be useful to add?
- the `nrnivmodl.cmake` file is installed alongside NEURON
- the `nrnivmodl.in` script is `configure_file`-able; only one parameter
  has practical significance, and that is the @CORENEURON@ one (set to
an empty string if using in `nrnivmodl` mode, and to "CORENEURON" if
using in `nrnivmodl-core` mode). It first tries to copy the `nrnivmodl.cmake` file to
the current directory; if the current directory already has a
CMakeLists.txt that differs from that one, it errors out. If the two are
identical, it runs the equivalent of `cmake -B [ARCH] -DMODFILES="$@"`.
Depending on whether or not the script is called `nrnivmodl` or
`nrnivmodl-core`, it also adds the `-DCORENEURON=CORENEURON` argument.
The script also sets `CMAKE_PREFIX_PATH` to where NEURON is
installed (so it can find NEURON using the `neuronConfig.cmake` file),
but has the option of overriding this behavior (for purposes of
debugging).

TODO list:
- I still need to remove the original `nrnivmodl-core` because it is
  rather complicated (do we need so many options?)
- remove `nrnivmodl_makefile` and co. so we are finally free of
  hand-made makefiles