Fix building NMODL with NRN_PYTHON_DYNAMIC
NMODL was being built using pybind11 CMake functions, which are, much
like Python detection of `find_package`, somewhat limited when it comes
to multiple versions. Now we can build it using
`add_nrn_python_library`, which more or less has the same functionality,
but is more customizable when it comes to multiple Python versions.