CMake-ify `git2nrnversion_h.sh`, remove `nrnversion.sh` (#3473)
Some facts:

- the only purpose of `git2nrnversion_h.sh` is to write `nrnversion.h`
- `nrnversion.h` contains only macros, which are only visible to the
  preprocessor
- `nrnversion.h` is only included in `nrnversion.cpp`, which is part of the
  `nrniv_lib` target

Therefore, we can just add the macros from `nrnversion.h` directly to the
`nrniv_lib` target as compile definitions, without having to write the
`nrnversion.h` file in the first place.

Additionally, the `bldnrnmacpkg.sh` uses `nrnversion.sh`, whose sole purpose is
to read the `nrnversion.h` file. On the other hand, we can just use `git
describe` to get the latest tag in the history (on `release/8.2`, this outputs
`8.2.7`, which is exactly the name of the MacOS package), so `nrnversion.sh` is
not needed either.

Other changes:

* Throw error if MacOS bundler encounters a shallow repo