Fix: #3583 (#3584)
Python nrncore_arg was overriding cell-permute from file even if the value was not given. In fact, if the value is already set CLI11 does not apply the value (as it should be).

Solution: check that the value was set (and not taken from default) before, eventually, setting it

Fix: #3583